+++ MLBB/60FPS/OpenGL/mlbb.sh
@@ -4,7 +4,7 @@
 # Mobile Legends Bang Bang All Tweaks for Smartphone [root] and [non-root]
 # Name      : MLBB Tweaks Performance
 # Author    : Algorithm Dev / Kreapic (t.me/algorithmdev)
-# Version   : 1.0-Beta
+# Version   : 1.0.2-Beta
 
 # ========================================================================