+++ MLBB/60FPS/SkiaGL/mlbb.sh
@@ -4,12 +4,25 @@
 # Mobile Legends Bang Bang All Tweaks for Smartphone [root] and [non-root]
 # Name      : MLBB Tweaks Performance
 # Author    : Algorithm Dev / Kreapic (t.me/algorithmdev)
-# Version   : 1.0-Beta
+# Version   : 1.0.2-Beta
 
 # ========================================================================
 
-# Next Update SKIAGL Optz
+# Add [device_cogfig] configuration mode
+sleep 2
+device_config get game_overlay com.mobile.legends
+echo " 🕗 : Checking Game Intervension            [OK] "
+echo " "
 
-# ========================================================================
-
-# I need testing to ensure the previously written code is successful and does not interfere with system performance.
+# Set device_config to 60FPS
+# device_config put game_overlay <PACKAGE_NAME>
+# mode=2,downscaleFactor=0.9:mode=3,downscaleFactor=0.5
+# source : https://developer.android.com/games/gamemode/gamemode-interventions
+device_config put game_overlay com.mobile.legends mode=2,skiagles=1,downscaleFactor=0.75,fps=60:mode=3,skiagles=0,downscaleFactor=0.75,fps=60
+echo " 🕗 : Throttling FPS Setting                [OK] "
+sleep 1
+echo " "
+# https://developer.android.com/games/gamemode/gamemode-interventions#fps_throttling
+echo " 🕗 : Configuration....                [SUCCESS] "
+sleep 1
+exit