+++ MLBB/90FPS/mlbb.sh
@@ -1,0 +1,12 @@
+# !/bin/sh
+# ========================================================================
+
+# Mobile Legends Bang Bang All Tweaks for Smartphone [root] and [non-root]
+# Name      : MLBB Tweaks Performance
+# Author    : Algorithm Dev / Kreapic (t.me/algorithmdev)
+# Version   : 1.0.2-Beta
+
+# ========================================================================
+
+# For next update i will added.
+# I need test performance device now